Gradle setup for Health Connect project.
The setup is inspired by:
- [SysUIGradleProject](https://source.corp.google.com/android/vendor/unbundled_google/packages/SystemUIGoogle/studio-dev/)
- [AndroidStudioPC](https://source.corp.google.com/android/vendor/google/apps/GooglePermissionController/studio-dev/)

This is the minimum necessary code to run an IDE with complete syntax highlighting and autocomplete for UI work.

Some elements will still need changing as we go along, for example the signature keys required to build the app directly from the Android Studio interface. For now please continue using `m` and `atest`.

This setup should ultimately be moved to the vendor/ package, but I am uploading it here for now in order to let the team test it.

Bug: b/255738373
Test: manually build project
Change-Id: Ieefc788c4c66b8f9b29b82f7769f08f5476ff80e