Implementation of the deletion flows.
The DeletionFragment uses a combination of Fragment Results APIs and a
shared viewModel to achieve reliable and readable Alert Dialog transitions.

The DeleteUseCase currently suspends for two seconds and randomly throws
an error to simulate the future API integration.

TODO: The remaining tests for the progress, success and failure dialogs
will be updated in a subsequent CL.

Screenshots:
    - Time range picker Dialog: https://screenshot.googleplex.com/6Yfgan68LUk5D3Y
    - Confirmation Dialog: https://screenshot.googleplex.com/7XFLtrNiedovgnj
    - Progress Dialog: https://screenshot.googleplex.com/AehcrnW2DUMBnHR
    - Success Dialog: https://screenshot.googleplex.com/9jZ9tjAGbWxuaGz
    - Failure Dialog: https://screenshot.googleplex.com/6FfssGeCUnNmeCU

Test: atest
Bug: b/245515048
Change-Id: Ib00fffec6e6cfe77c8bf6bc488bb0f033f341aa1