Make set/get of various HC states synchronized
We recognize that it's not possible to completely help with the below
situation:
Two processes get the states (synchronously) and find it fit to move
ahead with some operation. If they both try to do it at the same time
then one of them will fail and will have to retry. But it will fail
cleanly.

But this does help with the below situations:
- If one process is already in the middle of setting a state, the other
  should not be allowed to set that state at the same time. Otherwise we
  can get into a very messy state machine.
- If one process is already in the middle of setting a state, the other
  should wait for the current update to complete before reading.

Test: atest HealthConnectUnitTests:com.android.server.healthconnect.HealthConnectServiceImplTest
Test: atest CtsHealthConnectDeviceTestCases
Bug: 243145978
Change-Id: I6617f928e92709373fbeea935050339072a486f3