Stop closing HC DB
HC runs in a multithreaded environment and SQLite DB should not be
closed in such cases. This leads to unexpected SQLite exception
wherein one thread closes the DB while the other thread is still
working on an open instance of it. SQLite in such a cases closes the
DB for all the threads.

NOTE: HC's thread design is soon to be changed, hence this CL doesn't
add any multithreading tests. And relies on future CLs to do that.

Bug: 265340385, 265121971
Test: Tested using manual setup using atest
Change-Id: Ia8e35f45fc7e6de52310443350f896feafe7b544