Change thread model of HC
Introduce a new thread model for HC to use the following 3 thread pools,

- BACKGROUND: For internal background tasks
- SHARED: For all the client requests
- CONTROLLER: For all the controller APK requests

Further more for SHARED thread pool we now use a RR scheduling for the
taks for the clients so the system resources can be eually shared by the
clients in case there are multiple pending reuests from different
clients. This would prevent a single app from consuming all the
resources during a resoruce crunch.

The current number of threads in each executor are based on
the recommendation from system health team.

Bug: 265337296
Test: Manually + atest CtsHealthConnectDeviceTestCases

Change-Id: Idfba08d752ef17d2dbe04302c0279bf3abfc50bc