Send restored data to HC for final restoration
Once the Restore apk has finished restoring all HC data in the
BackupRestore app's dir, it sends all of that data over to the HC
service for final restoration.

Once the staging is finished, it deletes all the restored data from it's
own dir.

Add permission to allow access to HC data for d2d

This permission will be used to protect the access to HC data during a
d2d flow.
Only this app will be able to hold this permission as it's a signature
permission.

https://docs.google.com/document/d/183h6sa95LRxW3M6efyjjSLfVWXYdIyZUsTUmnnGh7W0/edit?resourcekey=0-bawquQwx075d7tk7jzeeYg#heading=h.68g4ihg96xkx

Test: atest HealthConnectBackupRestoreUnitTests
Bug: 243145975
Change-Id: Ie4c8e37c943ba59797ce3d18b990e779d980777d