Enable multi user support for HC
This CL adds support for adding health data on a per user basis.

Key points to note,
- HC only support queries on the current user
- This CL adds a per user DB, and the correct DB is based on the
current user id.
- Simply clearing caches is good enough for us, as we repopulate our
caches if our caches are empty

Note: Manually inspected that the correct DBs are being created,
and inspected the DB to see new required IDs are being generated as
well.

Bug: 260104124
Test: atest CtsHealthConnectDeviceTestCases on both multiple users

Change-Id: I045e522066228e903e10b02235ba886a86c826c8