Change to check if app is in foreground during readRecords and getChangeLogs API calls
read records and change logs should be allowed only for the foreground
apps. Whenever there is a read/chnage logs request  from an app this change adds a check
using AppOpsManagerLocal service public exposed API isUidInForeground() to check if the the app is in the foreground
and allows only if app is in foreground else throws a Security Exception
for background apps.

BUG: 265746621
Test: atest StepsRecordTes and manual
Change-Id: I80fab7a1540b63b764bf1cd1c7d64cf68a27f0d6