Check permission before any other condition
An app without the permission calling the API when not in foreground
should not be given the impression that it has the permission and that's
why why the only condition blocking it is that it's not in foreground.

That's why we have CTS tests that require Security exception is thrown
if the cts app doesn't have the permission.

This also brings consistency with the readRecords API which checks the
permission before checking foreground-ness.

Bug: 265746621
Bug: 269225908
Test: atest CtsHealthConnectDeviceTestCases
Change-Id: Ife98201938a472e1ec49b472a1113d8361af389e