Reset the thread pool in userswitching + improve code thread safety
For HC it might happen onUserSwitching was called and we paused our
scheduler in response but never started it back in onUserUnlocking
due to either the unlock is eventually happening on the same user.
The code gives us no guarantee that onUserUnlocking will be called
after onUserSwitching, or for other reasons. Either ways we can
pause and resume our threads in onUserSwitching which is what this
CL is doing.

This is root cause of b/268486060 - from logs

"HealthConnectScheduler: Unable to schedule task for uid: 10264"
This happens when the scheduler is paused.

Static non final vars in java are subject to race conditions,
and there is no gurantee that the a thread reads the latest value.

getInstance is inherently not thread safe right now

Bug: 268486060
Change-Id: I15e87ef80893a8435bc7c56ae3c916b51a3ef1eb