Change to throw an exception if both page token and different sort order are set for read request
Details of sort order for read request results is stored in pageToken.
Even pageToken reperesents results sorted in ascending order and odd
pageToken in descending order.
When both pageToken and sort order are set for a subsequent request, and the ordering
is changed in the request then an exception is thrown.

BUG: 266675570
Test: atest CtsHealthConnectDeviceTestCases
Change-Id: I7975d3e40f1a05d258ba3ef02ea45a2a606ad807