Don't back up large files for 25mb cloud backup
HealthConnectBackupAgent performs both the cloud backup / restore and
the d2d of health connect data. We don't want to include the large
health db file in the 25 mb cloud backup.
If the db is over 25 mb then BR will not backup anything for us.
Even if the health db is not large, we will not be backing it up.
Reasoning and decision:
https://docs.google.com/document/d/1C86-CTHz8NGFPp9KM8X-IDH9lv_BxHSf0k5hLJYgYcc/edit#bookmark=id.ms7n1mttjwaq

Test: manual
Test: atest HealthConnectBackupRestoreUnitTests
Bug: 243145978
Bug: 243145975
Change-Id: Iccb2dd7b156a0583f83656ba9f07a537ca78daba