Implement timeouts for various HC states
We don't want HC module to be stuck in a state forever (in case due to
some reason we are not able to get out of it per design of the state
machine).

After a reboot we check whether some timeout was scheduled, and if so
then we continue that.

Test: atest HealthFitnessUnitTests:com.android.server.healthconnect.backuprestore.BackupRestoreTest
Bug: 271078264
Change-Id: I51fe010babfd72be77488d357a18fcebbe764bfe