Revert "Revert "Executing BR state read and migration state update within read lock.""
As we are fetching BR state and then updating migration state, there
exists possibility of race condition. To avoid the race condition, we
are wrapping this in read lock(same lock used by BR).

The commit was reverted as the previous change and change in
MigrationStateManager.updateMigrationStateGuarded was merged at the same
time.

This reverts commit 7010dfe0af657764e11d852700403f57e6b301da.

Bug: 273266398
Test: atest CtsHealthFitnessDeviceTestCases
Change-Id: I2436026c86dc0465e69638560ffad1164429256d