Attempt merging of staged data on reboot
It's possible that the system rebooted while we were in the middle of
merging or about to begin merging. In such a case, we should attempt to
merge the staged data.

If we don't do this then due to an unknown system reboot we will not
perform merging and the user will not see their health data restored.

- Also, do the scheduling work on a background thread.

- Also, regrouped the methods per access modifiers.

- Added more tests.

Bug: 279896513
Bug: 279815142
Test: atest HealthFitnessUnitTests:com.android.server.healthconnect.backuprestore.BackupRestoreTest
Change-Id: Ie7d745323157c0a1deaa0559e24197f98956088d