Reset the grant times after merge cts test
testMergeStagedData_withEmptyHealthDb_mergesAllData merges the grant
time for the cts app and brings it to an old date.
Some other cts tests rely on current time to be the grant time, and an
older grant time breaks them.
So, resetting the grant time to current time once this test is over.

Tested by reproducing and then fixing.

Bug: 281853028
Test: atest CtsHealthFitnessDeviceTestCases
Change-Id: Id60a8e9233c035d015373db4bd9fa5aaaca1d1b4