Move merge test to integration tests
testMergeStagedData_withEmptyHealthDb_mergesAllData doesn't test any api
behavior, and tests the merging part of the restored data. So, this is
an integration test.

MUST_SLEEP: wait for the merge to complete; other synchronization method
like CountDownLatch doesn't work as the test won't get any trigger to
countdown, and so the await would be equivalent to sleep and won't be an
intended use-case for CountDownLatch.

Bug: 283751591
Test: atest HealthFitnessIntegrationTests
Test: atest CtsHealthFitnessDeviceTestCases
Change-Id: Ic15dc9563e40357de3fadac1de0c32dd4aa9ad47