Update the user for staged db context
Otherwise the staged db always works with the primary user's db.
This is not an issue right now, as backup-restore is supported for the
primary user only. But in the future if backup-restore is supported for
secondary users as well, then this will be an issue.

Tested with secondary user on pixel.

Bug: 281099471
Test: atest HealthFitnessIntegrationTests
Change-Id: I2cefc23a96b2329d0528d2ac3d7d4df63ce9a2e6