Add null checks for all public APIs in HealthConnectServiceImpl.
These null checks would result in the error being returned to the caller
in the Parcel. The existing behaviour was an NPE which was caught in a
catch-all in the HCThreadScheduler.

Also added null checks for some unused params in case we start using
them later, but this has a low chance of resulting in a behavious
change.

Added @NonNull annotations to arguments which were effectively
non-nullable but missing the annotations.

Fix: 288243721
Test: atest HealthFitnessUnitTests
Change-Id: I3f78e94d7d25cc3dbe0fe48aaf2c4afe76ef9eda