Temporary fix for failing CTS logging test.
The issue occurs when running the HealthConnectUiLogsTests on a fresh
install of Health Connect so that the Onboarding page appears. The CTS test setup installs test apps which insert data into Health Connect, so they appear in Recent Access. The permissions for these apps are granted as part of the setup so they do not need the UI to request permissions. Therefore, the first time Health Connect is opened during this test the Onboarding activity will be started.

Due to our setup, MainActivity is first started, which inflates the HomePage fragment, which in turn starts logging elements. We then redirect to the onboarding activity, the logging page ID is set to ONBOARDING_PAGE but the recent access elements from the HomePage still get logged, this time under the wrong page name (ONBOARDING_PAGE). The long term solution is probably to have a Gateway activity to redirect to Onboarding/Migration activities and avoid such logging issues.

Test: atest
Fix: 290048248
Change-Id: I2bb38db1064c51b1e6c858b0bd45beaf1b8b7ece