Retry getContributorApplicationsInfo call
Information about contributing apps is updated asynchronously outside
insert calls, leading to flakiness in the test. There is no reasonable
way from the test to wait for the update to complete, so instead we
retry a finite number of times with a small delay.

MUST_SLEEP Using Thread.sleep to delay between retries.

Test: atest
Bug: 288323923
Change-Id: Ibfed23c625de4c894b044b271ebaea7bb1cf28e8