Fix system crash that happens when making multiple API calls.
This crash happens because the JNI reference table hits its limit, if
multiple API calls are made after exceeding the API quota.

When the API quota is exceeded finishDataDelivery is not called.

This change moves the API quota before data delivery is started so no
references are allocated to the JNI table.

Fix: 301045666
Test: Tested locally using CTS tests in
https://googleplex-android-review.git.corp.google.com/c/platform/packages/modules/HealthFitness/+/24890767.

Change-Id: I158a28175b4d57df6e87181cfa50ccb748a5b094