Fix flaky onboarding test
scenario.state method is deprecated in favor of the scenario.getState() method. The scenario.getState() method is more specific and provides more information about the current state of the activity lifecycle. It also takes into account the current state of the activity's fragments, which the scenario.state method does not.

Bug: 311384996
Test: atest
Change-Id: I80d85cb81a31eddba3637d5b8f5dbd6b0c0b11ee