Ensure app info exists for priority list entries
Otherwise we store -1 for the app id, which then results in an exception
when the priority list is read back. Controller behavior should avoid
this happening in practice, but it's annoying for tests where we have to
ensure that app info has been created before setting priority.

Bug: 385122370
Flag: EXEMPT bug fix
Test: atest HealthDataCategoryPriorityHelperTest
Change-Id: Ia736b734f9770c17cb34f180a7996a3c1cdbc68c