Launch onboarding activity when available
This introduces the onboarding activity flow. If the developer has
exported an activity that responds to android.health.connect.action.ACTION_SHOW_ONBOARDING
Health Connect will launch the exported activity when the user attempts
to connect the app. This gives the app an opportunity to carry out any
setup that is required to connect to Health Connect. After this has been
done, the app should make a permission request, completing the flow.

Bug: 270672374
Flag: com.android.healthfitness.flags.launch_onboarding_activity
Test: atest
Change-Id: I0fe66df55185d36e78a3ac5ef585a4ca02b2d067