Remove uses of ExtendedMockito from BackupRestoreTest.
The main way to accomplish this is to create a fake file of a DB with a given version, rather than fake the return value.

We close the database in this test because it is not the HealthConnect database, it is a database used to create a file.

While doing this, I removed the usage of eventually(). This was not great for a couple of reasons:
- it used code from the CTS test library in unit tests
- it worked via polling.
As we already owned the FakePreferenceManager, it made sense to do the concurrency properly and signal when the condition was reached.

IGNORE_HC_DB_WARNING=true

Bug: 396076977
Flag: TEST_ONLY
Test: atest HealthFitnessUnitTests
Change-Id: Ie498812030483f0a399990c1e115845ce87d72d7