Set default units based on system settings rather than hardcoded.
Also stops writing the default unit to the preferences when the user hasn't manually set the unit. This means that when the user changes the system units this will be used in Health Connect.

As the default unit was written to preferences, there's no way to distinguish between users that are using the default and users that have explicitly set the unit. This means users won't automatically start seeing the new locale based units and will have to manually change this in Health Connect.

Bug: 381385712
Flag: EXEMPT bug fix
Test: atest
Change-Id: Ia01579f08db117373c0e1432243cbb91732f87d9