Fix ParcelUtils so it is doesn't make assumptions about Marshalling.
ParcelUtils was making an assumption that the size of a parcel when marshalled was equal to dataSize().
This looks reasonable, and fits with the default implementation. However, the CTS tests give no guarantee it is true.
In fact, Parcel can choose what it wants to use for its marshall output format. And in fact Robolectric makes a different choice.

Having had some discussions with the Robolectric team I'm convinced this assumption is wrong of us, and the correct behaviour is to use the actual marshalled size.

Bug: b/403609828
Test: atest HealthFitnessRoboUnitTests
Flag: EXEMPT in practice the code will act no differently in default android

Change-Id: Iaa7da45e30698447c83766b26abee0bd3a42699c