Change AggregationRecordData to an immutable design and clean up some NullAway warnings.
This change started as just addressing the NullAway issues.
But while trying to fix those, it became clear that moving the design to an immutable design where values were set in the constructor had some big advantages.
The point at which values were set became clearer.
It avoided the need for some unnecessary values in tests.
And it avoided the need to store database column names after they were needed.
I think the result is cleaner and safer.

Bug: b/317029272
Flag: EXEMPT pure refactor
Test: atest HealthFitnessUnitTests
Change-Id: Ibbc04ab597f7ae79ab43c554175bc28c51a137d1