Increase the thread safety of the device cache.
This started with me trying to just add Nullness annotations. But then I realised there was a possible race condition.
If the cache was cleared in between populateDeviceInfoMap being called, and returning the value the methods could return null (thank you NullAway).
So I've introduced a record as a pair type, and hopefully increased the null safety and removed the race condition.

Bug: b/317029272
Flag: EXEMPT annotation only
Test: atest HealthFitnessUnitTests
Change-Id: I1c77275ccb4e8f6fd22f64e266d7cf929d1f7b19