Make PriorityMigrationPayload null safe.
This is a slightly harder CL than some of the others removing NullAway annotations.
As this is a parcel coming in, we can't protect the input from potentially being null.
But DataOrigin is labeled as not having a null package name. So we have to decide what to do with null input.
In this case skipping seems appropriate - adding a DataOrigin with no information to the list adds nothing.
All of the code which uses it checks for the presence of a package name, so this looks safe.

Bug: b/317029272
Test: atest HealthFitnessUnitTests
Flag: EXEMPT no user visible impact
Change-Id: I6fc6e5c662ab4df6897818a3bbe45285c406c7d2