Set a default empty value for samples in some internal records.
I've checked by code inspection that this value should never be used.
However, with the code structured as it is at the moment, as a builder where the values can be set later, the only way to ensure the return value for samples is non null is to set a default value.

Bug: b/317029272
Test: atest HealthFitnessUnitTests
Flag: EXEMPT invisible to user
Change-Id: Iff207b1f058537315d338059bfd6eba1e230d853