Mark some HealthConnectService methods as nullable in the aidl and clear up suppressions in HealthConnectManager.
Note this is more complicated than you might think.
THere exists @nullable for AIDL https://source.android.com/docs/core/architecture/aidl/aidl-annotations#nullable
However, the docs for this say "This annotation is no-op for the Java backend. This is because, in Java, all non-primitive types are passed by reference, which could be null."
Now, this is definitely true. However, it is also true for methods in Java, even ones marked @NonNull. The annotation just helps the compiler.
So you could argue that @nullable ought to add an annotation for Java. This could be done in system/tools/aidl/generate_java.cpp
However, for now the JavaPassthrough works fine with NullAway and ErrorProne, and doesn't require AIDL changes.

Test: atest HealthFitnessUnitTests
Flag: EXEMPT annotation only
Bug: b/317029272

Change-Id: I2f660164bdd63b9b8ad89d22a62e0aef225a7f3d