Move post upsert/delete tasks to within the FitnessHelpers.
This should reduce the chances of bugs when we use the internal APIs to
insert data, like with cloud restore and native step tracking, as it
should make it difficult to miss these steps.

There are existing users of the internal APIs (d2d, import), which are not doing the
post upsert/delete steps today. In some cases, these flows take care of
calling the post tasks once at the end (and in some cases, not). I
haven't updated these flows in this change yet to keep this no-op.

There is a small behaviour change in that the post tasks are now being
scheduled before we return the result. But I don't expect this to have
any meaningful impact, since they are only being scheduled and not being
executed before.

Test: presubmit
Flag: NONE small change, see above.
Bug: 382009199
Change-Id: I22f15f685530b74fc21c7d8ca781a65169e1d97b