Make AggregationTimestamp null safe.
It is clear from the design that mParentTimtestamp is null for GROUP_BORDER records but not for INTERVAL_START or INTERVAL_END records.
So mParentRecord and getParentRecord() have been marked nullable.
Uses of getParent() have been checked for null safety, and then Objects.requireNonNull() have been added, and comments where necessary to those uses which are null safe.

Test: atest HealthFitnessUnitTests
Bug: b/317029272
Flag: EXEMPT On standard paths there is no user visible change.
Change-Id: I338c77ec68f28739ead985a90ec5dcf13bbedbfd