Free scrolling of contextual prompt when requesting multiple health permissions
Before this CL, we are using a "Show all" button, so that by default all health permissions are hidden, and users need to click to see granular data types.
With this CL, users are able to free scroll by default without having to click "Show all" button to expand.

Bug: 408122501
Flag: EXEMPT bugfix
Change-Id: I09556a48f794d7e6e42484b0ea311af93bd4026a