Add rate of perceived exertion to ExerciseSessionRecord
This is in addition to the new fields added to ExerciseSegment (which also includes RPE), so RPE can be found in both places.

The APIs in ExerciseSegment are currently marked with a hide annotation so they can be sent for API review together with the other segment changes, and so adding CTS tests is not yet possible, so this will be added when they are unhidden.
Bug: 414602798
Test: unit tests
Flag: com.android.healthfitness.flags.exercise_segment_improvements

Change-Id: Ie9ff521f08be8f25d0009a60762f372f8a3031f3