Try to insert AppInfo for access logs
This goes beyond the original fix in ag/29464526 by attempting to insert
an AppInfo when recording an access log for an app that we haven't yet
inserted an AppInfo. If we fail to insert the AppInfo we still silently
drop the access log. I can see an argument for failing the request
entirely, but not doing that yet as it would be a riskier change.

This is more robust, as we're no longer relying on something else to
have inserted the AppInfo, and simplifies testing.

Bug: 367656073
Bug: 381228944
Flag: EXEMPT bug fix
Test: atest HealthFitnessUnitTests CtsHealthFitnessDeviceTestCases
Change-Id: I211e90e05f11b885d611a3ea3e5464c554d488d9