Create an executor to process passive tracking sensor events.
When we receive a new step event, if it has been over 60 seconds since the last flush, we'll write all of the steps to the HC database. With this approach and given that we aren't implementing on demand flushing, it could be possible for a flush to take place, a user walks for 50 seconds and then stops and then those 50 seconds of steps are not written (as no steps took place over 60 seconds after the last flush). To resolve this, when we receive a step event, we can schedule a task 60 seconds in the future to flush to the database. If we receive more steps within 60 seconds, we'll push the scheduled task to 60 seconds in the future again. This will ensure that steps are always written in a timely manner and there should be minimal battery impact as the scheduled task only runs once after a user stops taking steps. I may also use this executor to process events from `onSensorChanged` as it may be called very frequently on the main thread.

Change-Id: Ia1e17e4d71a2d9a110d0c6c9e10c72a55c546557
Bug: 414765467
Flag: com.android.healthfitness.flags.step_tracking_enabled
Test: unit tests