Take Notification instead of notification type in HCNotificationSender
In current implementation, the same Int value passed to
`sendNotificationAsUser` means different things depending on the
notification factory supplied. This can be error prone.

Also the parameter is not annotated correctly.

In this change the method takes a Notification instance, so there's no
confusion and we can annotate the factory method with proper annotation.

Bug: 414949807
Test: atest HealthFitnessUnitTests
Flag: EXEMPT refactor
Change-Id: I93ae28e63a8b7a11d752160988aa03300bd4cb46