Create composables for each formatted classes.
When a record is formatted, it returns a formatted class from the parent
class `FormattedEntry`.
The formatted class holds the data to be displayed in a common
structure.
the formatted class is then passed onto `mapEntryToComposable` which
then returns it's corresponding composable function to display to the
user.

Formatters are yet to be submitted but i have included StepsFormatter in
this CL for a full example of the process.

Bug: 395934080
Bug: 415323998
Flag: NONE testapp
Test: atest

Change-Id: I4aeefe13d20841b08c15c391766cdd9bf713eb4e