Add batching and scheduled flushes to the native step tracker.
We are keeping track of the following fields:
- `mLastSavedData` which is the last data written into the HC database
- `mPendingData` which holds steps due to be written into the HC database
- `mPendingBatchWriteFuture` which is the scheduled future that will write any `mPendingData` into the HC database

After device boot or after a period of inactivity, when we receive a step event, we'll immediately write that step into the database. At the same time, we'll schedule a future to run in 60 seconds. If there are more step events within those 60 seconds, when the scheduled future runs, we'll write those steps into the database and schedule another future to run in 60 seconds. We'll keep doing this until there is a period of 60 seconds where no step events were received.

Bug: 397400522
Flag: com.android.healthfitness.flags.step_tracking_enabled
Test: unit tests
Change-Id: Ifdcf6d39fd3a79bf4ea38197fc48105d5d251242