Resolved TODO(b/376645901) in testMerge_doesNotCopyMedicalDataSourceDuplicates()
Regarding "TODO(b/376645901): Verify that there's no timeout with large datasets
on actual devices" - Verified on ab/13474601. The steps were:

- Used toolbox app to seed 750_000 medical resources
- Exported it to Google Drive, the file was 135MB.
- Deleted HC db file with `adb shell rm -r /data/system_ce/0/healthconnect/;`
- Imported again, took ~5 mins.
- Checked the number of resources via sqlite3 and `select count(*) from medical_resource_table;`, that printed out 750_000.

That certifies that the code path works, even though the UI can't show all the resources due to not having paging, that's a separate known issue to the UI team.

Bug: 376645901
Flag: TEST_ONLY
Test: atest
Change-Id: I7d9871a64080ec63f1a22101e729cfddf81589e3