Remove static instants from tests
Ideally the whole DataFactory should be a non-static class, or the methods in it that call now() should instead accept the time as an argument, but both of those options require a much larger refactor so this should do for now.

We still need to use the actual current time because a fixed timestamp doesn't work with historical read restrictions without extensive refactoring of tests.

Bug: 414434011
Flag: TEST_ONLY
Test: Yes.
Change-Id: I8f2c72800dda7e7113703ca1d84e67eee3cb1ea5