Fix auto deletion settings for backup & restore
The wrong preference key was used when backing up the auto delete settings so nothing was actually backed up. Also the stored auto deletion period is measured in days, this CL introduces a new string value and remove the previous enum proto. It is way harder to fix the issues for other unit preferences which might need to be removed in a follow up change.

Bug: 413688998
Test: manual and unit tests
Flag: com.android.healthfitness.flags.cloud_backup_and_restore
Change-Id: I275daeb24a1bb20f447f5e359d8c8f87a9b20217