Extrapolate the start time for step events.
We only receive steps when the AP is awake. If the device is asleep and the user takes steps, we won't know the start timestamp of the steps event. If we notice a small number of steps over a large duration (low cadence), we can trim and estimate the start timestamp by applying an estimated cadence. If the user is walking continuously (which triggers writes though delayed tasks), we'll use the end of the last event as the start for this one (meaning the duration of the event should roughly equal the batching period of one minute). Otherwise, if we notice a high cadence, we'll use the end of the last event as the start for this one.

Bug: 418989797
Flag: com.android.healthfitness.flags.step_tracking_enabled
Test: unit tests
Change-Id: Ica7aa3899158685dbfbda8515d6069e3a0ace57b