Remove write lock from updateOnboardingState method
If the new state is the same as the old state, write lock will be
acquired but never released (early return).

It's safe to remove because the write lock is also acquired at a
later time when we actually try to update the preference in
`updateOnboardingStatePreference`.

Fix: 421142899
Test: manual
Flag: com.android.healthfitness.flags.onboarding
Change-Id: I398bdadc7acb725289eb33310d673dcac3a2a7ea