Fix health connect crashes when browsing recent access
The crash occurs when the "Manage Permissions" button is clicked from the "Connected Apps" screen under specific circumstances.This issue is triggered if the user begins a swipe-to-go-back gesture from the right edge of the screen but cancels the action, causing the "Manage Permissions" button to erroneously appear. The button's click handler is not intended to be activated from this screen, leading to the crash.

Screen Recording: https://photos.app.goo.gl/sH1fzbTijBqMUgYT6

Bug: 421885566
Flag: NONE Bug fix
Test: RecentAccessFragmentTest
Change-Id: If437b6e20861820e9ec709952f75cd392c5f89aa