Convert some usages of nanos stored in longs to Instants.
This will allow the code to function beyond the year 2262 (which is the
max date in nanos that can be stored in a long before it overflows).

Sensor timestamps are since device boot and won't overflow until a
device is powered on for over 290 years so these will stay as longs.

Flag: com.android.healthfitness.flags.step_tracking_enabled
Test: unit tests
Bug: 418988134
Change-Id: I5677aaa51e811602173df24455214025a1608b83