Prevent device package from being repeatedly added to priority list
Previously, the DDP package was being added on every tracker refresh
(triggered on permission changes typically). Now, we only do it once,
the first time the tracker is initialized.

I've also added a test to the priority helper to verify that removal
of the DDP package is indeed possible when the set priority method is
called.

Test: unit tests
Flag: com.android.healthfitness.flags.step_tracking_enabled
Fix: 427733288
Change-Id: I218f701a8ce9f62e7a4470b611cc7d16630e8e81