Don't count steps from before tracking is enabled.
When step tracking is enabled, the sensor listener now establishes a baseline for the step counter. We flush the sensor so that the baseline can be set ASAP as it is based on the first sensor event received. This prevents steps taken before tracking was enabled from being counted (e.g. after user switch or if tracking permissions were granted for an app a while after device boot).

Bug: 427451398
Flag: com.android.healthfitness.flags.step_tracking_enabled
Test: unit tests
Change-Id: I3ae2e6ded47c93a31dc6a55937d07b563a6afc43