Add screen for managing device tracking
This adds a screen that will eventually allow the user to turn on/off
step tracking. Currently, it isn't wired up to the service, as the APIs
to control this are not ready yet. For now, the switch is just disabled
in the 'on' state.

Fix:427175257
Flag: com.android.healthfitness.flags.step_tracking_enabled
Test: DeviceManagementFragmentTest

Change-Id: Ia283c7784b492f82b0fc3bc830bdf1a0afc531ed