Fix caching issue for DDP display name special case
Special casing is applied to the "android" package so that instead of
naively using the "Android System" name from PackageManager, we use the
device display name. Whenever the in-memory cache is first populated, we
apply this special casing. However, when the "android" package is first
added, after the cache has been populated, we add it directly to the
cache. In this scenario, we weren't correctly applying the special
casing, leading to the "Android System" name showing up in various
places. This issue would disappear after a reboot due to the cache
getting repopulated.

This fix ensures that we apply the special casing when we add it
directly to the map, after initial population.

Fix: 431125379
Test: AppInfoHelperTest
Flag: com.android.healthfitness.flags.step_tracking_enabled
Change-Id: I06e6072f1bfce783498f33d0653324b364af3e18