Add stats collector for native tracking
This runs as part of the daily job and gathers various metrics related
to the native step tracker. Most of the stats can be derived from ad-hoc
DB reads, but the enablement/disablement status of the tracker requires
interacting directly with it.

I've omitted the logic to map failures/errors to an error code in this
CL and will tackle is separately, as it's a fairly significant change.
For now ERROR_UNSPECIFIED is set as a placeholder.

Bug: 416455504
Test: NativeTrackingStatsCollectorTest
Flag: com.android.healthfitness.flags.step_tracking_enabled

Change-Id: I6aa870e413b8b94c378174bd8d856ea1abc8158e