Only revoke necessary subset of permissions
Revoking all permissions leads to a race condition where the test app
may be killed after we started interacting with it again. The existing
solution/workaround of waiting for the app to be killed is insufficient,
as revoking many permissions can lead to more than one asynchronous
request for the app to be killed.

Fix: 433587449
Fix: 435045584
Bug: 439491034
Flag: TEST_ONLY
Test: atest CtsHealthConnectHostSideDeviceTestCases
Change-Id: I031f6d4c5a7f568e9c36dfe29c86a82682a468fe