Fix excess memory usage when bulk inserting
For large records, eg. sessions with routes, each UpsertTableRequest can
be considerable in size. Instead of constructing the requests for all
records to be inserted up front, we can instead do it on the fly inside
the transaction. This is functionally equivalent but avoids the memory
overhead.

This looks to make RecordUpsertTableRequest redundant, we can clean that
up in a follow up.

Bug: 443928914
Bug: 444206694
Flag: EXEMPT BUGFIX
Test: atest CtsHealthFitnessDeviceTestCases HealthFitnessIntegrationBackupRestoreTests
Change-Id: Ib69ede609833b3153b0396177f0f1c7b8a8c6b76