Add default visibility for DDP matchmaking flag
This change allows the DDP to specify whether or not a device should
show up by default in matchmaking. If true, the device will show up
alongside normal apps. If false (default), a developer must specifically
request the device package name to trigger matchmaking for it.

This gives the freedom to choose whether or not it makes sense for the
device data source to show up in matchmaking. In some cases, e.g. Wear
direct writes, it may be desirable to only show the watch in matchmaking
under certain conditions.

Fix: 445629466
Flag: com.android.healthfitness.flags.device_data_providers_api
Test: atest
Change-Id: I70b27dbff4e4ee5d720dd9ddc0d557c569c4d43d