Improve code quality and remove redundant observer.
On toggle the Data Category switch updates all permissions for that category. There is a global observer for individual permission states (not data category rather individual permissions) and on change in state for a permission the observer toggles switches for those permission preferences.

This logic existed in FitnessAppFragment and SettingsFitnessAppFragment.  It was replicated in the the FitnessPermissionsFragments.

This made the grantedFitnessCategories redundant.

Test: atest
Bug: 436747979
Flag: com.android.healthfitness.flags.permissions_grouping_ui
Change-Id: I0e8711a742ecb600d5b2939ee9327e4213333d56