Implement permission checking for device data providers
This verifies that the calling package holds
android.permission.PROVIDE_HEALTH_CONNECT_DEVICE_DATA. However, since
this is a privileged permission we need to introduce an alternative for
Android versions that predate the existence of this new framework permission.

The solution is to use config_systemActivityRecognizer which allows
packages in that list to act as device data providers.

Additionally, to support testing on Android versions that predate the
permission, we permit holders of MANAGE_HEALTH_DATA to act as device
data providers. This permission is held by the shell since Android U, so
that allows us to interact with the new APIs by adopting the shell
permission identity in test environments.

Fix: 437918245
Bug: 315116545
Flag: com.android.healthfitness.flags.device_data_providers_api
Test: atest DeviceDataProviderHelperTest
Change-Id: Idde8f443b18bf520f6fbe6e33aaf0191d35e6e84