Add public MatchmakingRequest and MatchmakingResponse classes
Before: The public isMatchmakingPossible() API accepted a Set and returned a boolean. The underlying, hidden GetMatchingAppsRequest was populated using this Set.
After: The public isMatchmakingPossible() API now accepts a MatchmakingRequest and returns a MatchmakingResponse. The callingPackageName field is hidden within the request, as it's only populated by the hidden getMatchingApps() API.

This refactoring was needed to prepare for future field additions to both MatchmakingRequest and MatchmakingResponse, ensuring both the request and result objects are properly wrapped in dedicated classes.

Bug: 452392975
Bug 416218092
Flag: com.android.healthfitness.flags.matchmaking
Test: atest

Change-Id: If7afb69c80d391eb28cedba4810e84499cb09428